+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All notable changes to this project will be documented in this file.
 
 ### Deprecated
 
+* Deprecated `AbstractDomainRepository->create_identity_model()`. Use `AbstractDomainRepository->createIdentityModel()` instead.
+* Deprecated `AbstractDomainRepository->find_all()`. Use `AbstractDomainRepository->findAll()` instead.
+* Deprecated `AbstractDomainRepository->find_by_id()`. Use `AbstractDomainRepository->findById()` instead.
+* Deprecated `AbstractDomainRepository->query_many()`. Use `AbstractDomainRepository->queryMany()` instead.
+* Deprecated `AbstractDomainRepository->query_one()`. Use `AbstractDomainRepository->queryOne()` instead.
 * Deprecated `AbstractModelCollection->get_by_id()`. Use `AbstractModelCollection->getById()` instead.
 * Deprecated `AbstractModelCollection->remove_item()`. Use `AbstractModelCollection->remove()` instead.
 * Deprecated `AbstractModelCollection->set_item()`. Use `AbstractModelCollection->add()` instead.